requirements.txt
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
idna==3.10
jplephem==2.22
mypy-extensions==1.0.0
numpy==2.2.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.3
sgp4==2.23
skyfield==1.49
urllib3==2.2.3
